% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DengueForecast.R
\name{dengue.forecast}
\alias{dengue.forecast}
\title{Dengue Forecast}
\usage{
dengue.forecast(h)
}
\arguments{
\item{h}{forecast horizon in weeks (a number greater than 0)}
}
\examples{
dengue.forecast(10) # Returns forecasts for the next 10 weeks
}
